+++ changelog.md
@@ -1,4 +1,16 @@
 ___
+## v2.12
+```
+DATE: 2023_11
+```
+
++ Add the **Human Orthologs** for the species: 'rnorvegicus','cgriseus','sscrofa','ocuniculus','drerio','ggallus','btaurus'.
+
++ Add the label of Principal Isoform from APPRIS database, the TRIFID scores and CORSAIR score for the species: 'rat
+
++ The addition of the cRAP database has been disabled for the moment.
+
+___
 ## v2.11
 ```
 DATE: 2023_06